softgpu: Reduce some minor bin item field sizes.

+++ GPU/Software/BinManager.cpp
@@ -166,7 +166,7 @@ void BinManager::UpdateState(bool throughMode) {
 	if (HasDirty(SoftDirty::PIXEL_ALL | SoftDirty::SAMPLER_ALL | SoftDirty::RAST_ALL)) {
 		if (states_.Full())
 			Flush("states");
-		stateIndex_ = (int)states_.Push(RasterizerState());
+		stateIndex_ = (uint16_t)states_.Push(RasterizerState());
 		ComputeRasterizerState(&states_[stateIndex_], throughMode);
 		states_[stateIndex_].samplerID.cached.clut = cluts_[clutIndex_].readable;
 
@@ -312,7 +312,7 @@ void BinManager::UpdateClut(const void *src) {
 	PROFILE_THIS_SCOPE("bin_clut");
 	if (cluts_.Full())
 		Flush("cluts");
-	clutIndex_ = (int)cluts_.Push(BinClut());
+	clutIndex_ = (uint16_t)cluts_.Push(BinClut());
 	memcpy(cluts_[clutIndex_].readable, src, sizeof(BinClut));
 }
